[ANDROSDK-1947] refactor network scripts names
